pyproject.toml
[tool.poetry]
name = "Annual-Report-Maker"
version = "0.1.0"
description = ""
authors = ["Pravin <PravinCoder@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
python-dotenv = "1.0.0"
crewai-tools = "^0.4.26"
embedchain = {extras = ["youtube"], version = "^0.1.116"}
langchain-anthropic = "^0.1.20"
ollama = "^0.2.1"
crewai = "^0.41.1"
langchain-groq = "^0.1.9"
flask = "^3.0.3"
pypdf2 = "^3.0.1"
python-multipart = "^0.0.9"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"